pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omnixas"
version = "3.0.0"
description = "Machine learning-based X-ray Absorption Spectroscopy (XAS) prediction"
requires-python = ">=3.12.4"
authors = [{ name = "Shubha Raj Kharel", email = "your.email@example.com" }]
dependencies = [
    "loguru>=0.7.2",
    "pip>=24.3",
    "ply>=3.6",
    "torch>=2.2.0",
    "torchvision",
    "torchaudio",
    "lightning",
    "hydra-core==1.3.2",
    "hydra-colorlog>=1.2.0",
    "omegaconf>=2.3.0",
    "optuna>=3.6.1",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.2",
    "numpy>=1.22",
    "matplotlib>=3.9.0",
    "seaborn==0.13.2",
    "tensorboard>=2.13",
    "plotly>=5.23.0",
    "pillow>=10.4.0",
    "tqdm>=4.66.4",
    "python-dotenv>=1.0.1",
    "dgl",
    "pymatgen>=2023.11.12",
    "matgl==0.8.5",
    "torchmetrics>=1.4.2",
    "torchsummary>=1.5.1",
    "python-box>=7.2.0",
    "pydantic>=2.9.2",
    "pyyaml",
]


[project.optional-dependencies]
dev = ["pytest", "black", "flake8"]

# Keep your existing tool configurations
[tool.black]
line-length = 88

# [tool.ruff]
# ignore = ["F401", "F811", "E402"]
# exclude = [
#     ".bzr",
#     ".direnv",
#     ".eggs",
#     ".git",
#     ".git-rewrite",
#     ".hg",
#     ".ipynb_checkpoints",
#     ".mypy_cache",
#     ".nox",
#     ".pants.d",
#     ".pyenv",
#     ".pytest_cache",
#     ".pytype",
#     ".ruff_cache",
#     ".svn",
#     ".tox",
#     ".venv",
#     ".vscode",
#     "__pypackages__",
#     "_build",
#     "buck-out",
#     "build",
#     "dist",
#     "node_modules",
#     "site-packages",
#     "venv",
# ]


[tool.setuptools]
packages.find.include = ["omnixas*"]
packages.find.exclude = ["*._legacy*", "*.tests*"]

# Include your config files in the package
[tool.setuptools.package-data]
omnixas = ["config/*.yaml", "config/**/*.yaml"]

# Add entry points for CLI if needed
[project.scripts]
omnixas = "omnixas.predict:predict_cli"

# Optional: Add additional project metadata
[project.urls]
"Homepage" = "https://github.com/yourusername/omnixas"
"Bug Tracker" = "https://github.com/yourusername/omnixas/issues"
